Add homework 4
