+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.mvc.view.prefix=/WEB-INF/view/
+spring.mvc.view.suffix=.jsp
